requirements.txt
fastapi~=0.117.1
uvicorn[standard]~=0.37.0
sqlalchemy~=2.0.43
python-dotenv~=1.1.1
passlib[bcrypt]~=1.7.4
python-jose[cryptography]~=3.5.0
python-multipart~=0.0.20
alembic
psycopg2-binary
email-validator~=2.3.0
pdfminer.six==20231228
openai
numpy
h11~=0.16.0
pip~=25.2
cryptography~=46.0.1
PyYAML~=6.0.3
sniffio~=1.3.1
typing_extensions~=4.15.0
click~=8.3.0
starlette~=0.48.0
pydantic~=2.11.9
pydantic_core~=2.33.2
websockets~=15.0.1
anyio~=4.11.0
httptools~=0.6.4
watchfiles~=1.1.0
annotated-types~=0.7.0
pymysql
wheel~=0.45.1

# Resume parsing dependencies
pyresparser
spacy
nltk

# After installing spacy, run the following separately:
# python -m spacy download en_core_web_sm
